Updates for differentail drive
I wrote two functions. Pivot and move straigh. I also changed some of the intialization stuff to reflect that we are talking about the Right and Left wheels instead of the X and Y axis wheels.  Havent tested it but I figured it was an aliright place to start. Eventually how we call these functions will be based upon our converstion of x and y coordinates into movement. I think we need to read up more on existing research before doing that part